--- ProyectoCompra/Ficheros/datosProvisionales.text
@@ -1,1 +1,0 @@
-1;
