+++ ProyectoCompra/Formularios/FrmInicioSesion.resx
@@ -117,37 +117,27 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
-  <data name="btnMostrarPass.BackgroundImageLayout" type="System.Windows.Forms.ImageLayout, System.Windows.Forms">
-    <value>Center</value>
-  </data>
-  <data name="btnMostrarPass.FlatStyle" type="System.Windows.Forms.FlatStyle, System.Windows.Forms">
-    <value>Flat</value>
-  </data>
-  <data name="btnMostrarPass.ImeMode" type="System.Windows.Forms.ImeMode, System.Windows.Forms">
-    <value>NoControl</value>
-  </data>
   <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
-  <data name="btnMostrarPass.Location" type="System.Drawing.Point, System.Drawing">
-    <value>220, 124</value>
+  <data name="contrasenia.Location" type="System.Drawing.Point, System.Drawing">
+    <value>27, 124</value>
   </data>
-  <data name="btnMostrarPass.Size" type="System.Drawing.Size, System.Drawing">
-    <value>19, 22</value>
+  <data name="contrasenia.Size" type="System.Drawing.Size, System.Drawing">
+    <value>219, 30</value>
   </data>
   <assembly alias="mscorlib" name="mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
-  <data name="btnMostrarPass.TabIndex" type="System.Int32, mscorlib">
-    <value>9</value>
+  <data name="contrasenia.TabIndex" type="System.Int32, mscorlib">
+    <value>8</value>
   </data>
-  <data name="&gt;&gt;btnMostrarPass.Name" xml:space="preserve">
-    <value>btnMostrarPass</value>
+  <data name="&gt;&gt;contrasenia.Name" xml:space="preserve">
+    <value>contrasenia</value>
   </data>
-  <data name="&gt;&gt;btnMostrarPass.Type" xml:space="preserve">
-    <value>System.Windows.Forms.Button, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  <data name="&gt;&gt;contrasenia.Type" xml:space="preserve">
+    <value>ProyectoCompra.Controles.ctrlContrasenia, ProyectoCompra, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null</value>
   </data>
-  <data name="&gt;&gt;btnMostrarPass.Parent" xml:space="preserve">
+  <data name="&gt;&gt;contrasenia.Parent" xml:space="preserve">
     <value>groupBox1</value>
   </data>
-  <data name="&gt;&gt;btnMostrarPass.ZOrder" xml:space="preserve">
+  <data name="&gt;&gt;contrasenia.ZOrder" xml:space="preserve">
     <value>0</value>
   </data>
   <data name="chxRecordar.AutoSize" type="System.Boolean, mscorlib">
@@ -156,11 +146,12 @@
   <data name="chxRecordar.Font" type="System.Drawing.Font, System.Drawing">
     <value>Arial Nova, 8.25pt</value>
   </data>
+  <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
   <data name="chxRecordar.ImeMode" type="System.Windows.Forms.ImeMode, System.Windows.Forms">
     <value>NoControl</value>
   </data>
   <data name="chxRecordar.Location" type="System.Drawing.Point, System.Drawing">
-    <value>152, 148</value>
+    <value>152, 154</value>
   </data>
   <data name="chxRecordar.Size" type="System.Drawing.Size, System.Drawing">
     <value>88, 18</value>
@@ -183,33 +174,6 @@
   <data name="&gt;&gt;chxRecordar.ZOrder" xml:space="preserve">
     <value>1</value>
   </data>
-  <data name="txtContrasenia.Font" type="System.Drawing.Font, System.Drawing">
-    <value>Arial Nova, 8.25pt</value>
-  </data>
-  <data name="txtContrasenia.Location" type="System.Drawing.Point, System.Drawing">
-    <value>27, 124</value>
-  </data>
-  <data name="txtContrasenia.Size" type="System.Drawing.Size, System.Drawing">
-    <value>187, 21</value>
-  </data>
-  <data name="txtContrasenia.TabIndex" type="System.Int32, mscorlib">
-    <value>6</value>
-  </data>
-  <data name="txtContrasenia.TextAlign" type="System.Windows.Forms.HorizontalAlignment, System.Windows.Forms">
-    <value>Center</value>
-  </data>
-  <data name="&gt;&gt;txtContrasenia.Name" xml:space="preserve">
-    <value>txtContrasenia</value>
-  </data>
-  <data name="&gt;&gt;txtContrasenia.Type" xml:space="preserve">
-    <value>System.Windows.Forms.TextBox, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-  </data>
-  <data name="&gt;&gt;txtContrasenia.Parent" xml:space="preserve">
-    <value>groupBox1</value>
-  </data>
-  <data name="&gt;&gt;txtContrasenia.ZOrder" xml:space="preserve">
-    <value>2</value>
-  </data>
   <data name="lblContrasenia.AutoSize" type="System.Boolean, mscorlib">
     <value>True</value>
   </data>
@@ -217,7 +181,7 @@
     <value>NoControl</value>
   </data>
   <data name="lblContrasenia.Location" type="System.Drawing.Point, System.Drawing">
-    <value>24, 106</value>
+    <value>24, 107</value>
   </data>
   <data name="lblContrasenia.Size" type="System.Drawing.Size, System.Drawing">
     <value>63, 14</value>
@@ -238,7 +202,7 @@
     <value>groupBox1</value>
   </data>
   <data name="&gt;&gt;lblContrasenia.ZOrder" xml:space="preserve">
-    <value>3</value>
+    <value>2</value>
   </data>
   <data name="txtUsuario.Font" type="System.Drawing.Font, System.Drawing">
     <value>Arial Nova, 8.25pt</value>
@@ -265,7 +229,7 @@
     <value>groupBox1</value>
   </data>
   <data name="&gt;&gt;txtUsuario.ZOrder" xml:space="preserve">
-    <value>4</value>
+    <value>3</value>
   </data>
   <data name="lblUsuario.AutoSize" type="System.Boolean, mscorlib">
     <value>True</value>
@@ -295,13 +259,13 @@
     <value>groupBox1</value>
   </data>
   <data name="&gt;&gt;lblUsuario.ZOrder" xml:space="preserve">
-    <value>5</value>
+    <value>4</value>
   </data>
   <data name="btnLogin.ImeMode" type="System.Windows.Forms.ImeMode, System.Windows.Forms">
     <value>NoControl</value>
   </data>
   <data name="btnLogin.Location" type="System.Drawing.Point, System.Drawing">
-    <value>27, 166</value>
+    <value>27, 172</value>
   </data>
   <data name="btnLogin.Size" type="System.Drawing.Size, System.Drawing">
     <value>212, 25</value>
@@ -322,7 +286,7 @@
     <value>groupBox1</value>
   </data>
   <data name="&gt;&gt;btnLogin.ZOrder" xml:space="preserve">
-    <value>6</value>
+    <value>5</value>
   </data>
   <data name="groupBox1.Location" type="System.Drawing.Point, System.Drawing">
     <value>81, 67</value>
@@ -1192,9 +1156,6 @@
   </data>
   <data name="$this.Font" type="System.Drawing.Font, System.Drawing">
     <value>Arial Nova Cond, 8.25pt, style=Bold</value>
-  </data>
-  <data name="$this.ImeMode" type="System.Windows.Forms.ImeMode, System.Windows.Forms">
-    <value>NoControl</value>
   </data>
   <data name="$this.StartPosition" type="System.Windows.Forms.FormStartPosition, System.Windows.Forms">
     <value>CenterScreen</value>
